requirements.txt
httpx
pydantic
pydantic-settings
python-dotenv
fastapi
uvicorn
watchdog
pytest
requests
python-uinput
